Horizontal collisions working to an extent.
I'll be honest, it didnt go as smooth as i thought it would, and i hope to god one of you guys have an idea.
It works, but it feels hacky and jerky.

We could fix the jerkyness by making the camera more smooth, some kind of lerp, rather than it moving exactly with the player.