Bug Fixed #005 - If the user jumps, then grabs a ladder by pressing up; they are no longer able to climb down the ladder
If the user jumps, then grabs a ladder by pressing up; they are no longer able to
climb down the ladder. They will either fall when pressing down or appear to be
standing in mid-air.

Video link:
http://www.youtube.com/watch?v=JaXrlqL-q_o

Fixed By adding a small impulse to the player to dislodge them. This is actually caused by a unity physics glitch :)